More index page refactoring
